+++ Front-End/src/pages/clubSearch/ClubSearch.jsx
@@ -1,11 +1,19 @@
-import { Container, IconButton } from '@mui/material';
+import { Container, IconButton, Box } from '@mui/material';
 import Categories from '../../components/categories/Categories';
 import Recomendations from '../../components/recomendations/Recomendations';
 import SearchBar from '../../components/ui/search/SearchBar';
 import Products from '../../components/products/Products';
 import SearchResults from '../../components/clubSearch/searchResults';
+import ArrowCircleLeftTwoToneIcon from "@mui/icons-material/ArrowCircleLeftTwoTone";
+import { useNavigate } from "react-router-dom";
 
 const Home = () => {
+
+  const navigate = useNavigate();
+
+  const handleClick = () => {
+    navigate("/");
+  };
 
   return (
     <Container 
@@ -13,16 +21,29 @@ const Home = () => {
       sx={{
         padding: '0',
         mt: '120px',
-        mb: '40px',
+        position: 'relative',
         display: 'flex',
         flexDirection: 'column',
         textAlign: 'center',
       }}
       id='home'
     >
+            <IconButton
+                aria-label="Volver"
+                color="#FFFFFF"
+                size="large"
+                onClick={handleClick}
+                sx={{
+                  position: 'absolute',
+                  top: 20,
+                  right: -90,
+                }}
+            >
+                <ArrowCircleLeftTwoToneIcon fontSize="large" color="#FFFFFF" />
+            </IconButton>
+            <Categories />
             <SearchResults />
             <Recomendations/>
-            <Products/>
         </Container>
     );
 }
